Move TimeFilter type to core
